增加对被订单引用的协议不能被删除处理、修改协议服务列表接口

+++ protocol-manage/src/main/java/com/zhiweicloud/protocolmanage/controller/ProtocolServController.java
@@ -8,6 +8,7 @@ import com.zhiweicloud.guest.APIUtil.PaginationResult;
 import com.zhiweicloud.guest.common.RequsetParams;
 import com.zhiweicloud.protocolmanage.model.ProtocolServ;
 import com.zhiweicloud.protocolmanage.service.ProtocolServService;
+import com.zhiweicloud.protocolmanage.service.ProtocolService;
 import io.swagger.annotations.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,6 +34,9 @@ public class ProtocolServController {
     private static final Logger logger = LoggerFactory.getLogger(ProtocolServController.class);
     @Autowired
     private ProtocolServService protocolServService;
+
+    @Autowired
+    private ProtocolService protocolService;
 
     @RequestMapping(value ="/protocol-serv-list")
     @ApiOperation(value = "协议服务管理 - 分页查询", notes = "返回分页结果", httpMethod = "GET", produces = "application/json")
@@ -158,6 +162,13 @@ public class ProtocolServController {
             @RequestParam(value = "airportCode", required = true) String airportCode) {
         try {
             List<Long> ids = params.getData();
+            boolean flame;
+            for(int i = 0; i < ids.size(); i++){
+                flame = protocolService.selectOrderByProtocolId(ids.get(i),airportCode);
+                if(flame == true){
+                    return LXResult.build(4998, "该协议被订单引用，协议下服务不能被删除");
+                }
+            }
             protocolServService.deleteById(ids,airportCode);
             return LXResult.success();
         } catch (Exception e) {
@@ -202,6 +213,9 @@ public class ProtocolServController {
             param.put("price",price);
             param.put("freeRetinueNum",freeRetinueNum);
             param.put("overStaffUnitPrice",overStaffUnitPrice);
+            if(protocolService.selectOrderByProtocolId(Long.getLong(protocolId),airportCode) == true){
+                return LXResult.build(4998, "该协议被订单引用，协议下服务不能被删除");
+            }
             protocolServService.deleteByType(param);
             return LXResult.success();
         } catch (Exception e) {
